added documents on the functions and modified vibronic_hamiltonian.py
